Simplify ParsingHelper, replace template example,
Change ParsingHelper to use a static method for reading the entirety of
a CSV file and returning an array, rather than accepting an interface 
to call the ProcessLine method (the PDE is not well suited to the 
latter approach).

Add the BasicDemo example and remove the original template example. The
new example relies on the new ParsingHelper.